refactor stylesheets in custom input, signIn, register, error feedback, create auth header to separate and reuse links signin/register
