+++ src/app/reset-password/page.tsx
@@ -1,4 +1,4 @@
-import ResetPasswordComponent from '@/components/ResetPassword/ResetPasswordComponent';
+import ResetPasswordComponent from '@/components/Auth/ResetPassword/ResetPasswordComponent';
 
 const ResetPasswordPage = () => {
   return <ResetPasswordComponent />;
