+++ src/components/Auth/AuthHeader/AuthHeader.module.scss
@@ -1,0 +1,19 @@
+@import '../../../sass/utils/container';
+@import '../../../sass/utils/mixins';
+
+.box__link {
+  @include flexCenter();
+  gap: 8px;
+  @include font(22px, 2vw, 25px, 400, 25px);
+}
+
+.link__sign_in {
+  text-decoration: underline;
+  text-decoration-color: var(--text-color);
+  text-decoration-thickness: 1px;
+}
+.link__register {
+  text-decoration: underline;
+  text-decoration-color: var(--text-color);
+  text-decoration-thickness: 1px;
+}
