+++ src/components/Auth/ErrorFeedback/errorFeedback.module.scss
@@ -1,0 +1,14 @@
+@import '../../../sass/utils/container';
+@import '../../../sass/utils/mixins';
+
+.error {
+  position: absolute;
+  bottom: -18px;
+  left: 0;
+  color: var(--invalid);
+  @include font(10px, 2vw, 11px, 400, 12px);
+
+  @include mediaQuery($screen-size-m) {
+    @include font(12px, 2vw, 14px, 400, 14px);
+  }
+}
