+++ src/components/Auth/Login/SignIn.tsx
@@ -1,43 +1,24 @@
 'use client';
 
 import Link from 'next/link';
-import { usePathname } from 'next/navigation';
 
 import { AppRouteEnum } from '@/libs/enums/enums';
 
 import SignInFormik from './SignInFormik';
 import SocialLoginGroup from '@/components/UI/SocialLoginGroup/SocialLoginGroup';
+import AuthHeader from '../AuthHeader/AuthHeader';
 
 import s from './signIn.module.scss';
+import common from '@/components/Auth/commonAuthStyles.module.scss';
 
 const SignIn = () => {
-  const pathname = usePathname();
-
   return (
-    <section className={`${s.section} `}>
-      <div className={`${s.container} `}>
-        <div className={s.box__link}>
-          <Link
-            href={AppRouteEnum.SIGN_IN}
-            className={` ${s.link} ${
-              pathname === `${AppRouteEnum.SIGN_IN}` ? s.link__sign_in : ''
-            } `}
-          >
-            Увійти
-          </Link>
-          <span>/</span>
-          <Link
-            href={AppRouteEnum.SIGN_UP}
-            className={` ${s.link} ${
-              pathname === `${AppRouteEnum.SIGN_UP}` ? s.link__register : ''
-            } `}
-          >
-            Зареєструватись
-          </Link>
-        </div>
+    <section className={`${common.section} `}>
+      <div className={`${common.container} `}>
+        <AuthHeader />
         <SignInFormik />
-        <div className={s.box__link__group}>
-          <p className={s.text}>Увійти за допомогою</p>
+        <div className={common.box__link__group}>
+          <p className={common.text}>Увійти за допомогою</p>
           <SocialLoginGroup />
           <Link
             href={AppRouteEnum.REQUEST_PASSWORD_RESET}
