+++ src/components/Auth/Login/signIn.module.scss
@@ -1,141 +1,7 @@
 @import '../../../sass/utils/container';
 @import '../../../sass/utils/mixins';
 
-.section {
-  padding: 40px 0;
-  @include mediaQuery($screen-size-s) {
-    padding: 74px 0;
-  }
-
-  @include mediaQuery($screen-size-l) {
-    padding: 94px 0;
-  }
-  @include mediaQuery($screen-size-xl) {
-    padding: 114px 0;
-  }
-}
-.box__link {
-  @include flexCenter();
-  gap: 8px;
-  @include font(22px, 2vw, 25px, 400, 25px);
-}
-
-.link__sign_in {
-  text-decoration: underline;
-  text-decoration-color: var(--text-color);
-  text-decoration-thickness: 1px;
-}
-.link__register {
-  text-decoration: underline;
-  text-decoration-color: var(--text-color);
-  text-decoration-thickness: 1px;
-}
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-
-  padding-top: 52px;
-  margin: 0 auto;
-
-  max-width: 400px;
-  @include mediaQuery($screen-size-s) {
-    gap: 35px;
-  }
-  @include mediaQuery($screen-size-m) {
-    max-width: 530px;
-    gap: 48px;
-  }
-
-  @include mediaQuery($screen-size-xl) {
-    max-width: 584px;
-  }
-}
-
-.form__box {
-  position: relative;
-}
-
-.label {
-  position: absolute;
-  color: var(--color);
-  top: -11px;
-  left: 11px;
-  background: var(--background-color);
-  padding: 0 18px 0 12px;
-  border-bottom-left-radius: 10px;
-  @include font(14px, 2vw, 18px, 400, 18px);
-  &.invalid {
-    color: var(--invalid);
-  }
-}
-.input {
-  width: 100%;
-  border: var(--border);
-  border-radius: var(--border-radius);
-  padding: 12px 40px 12px 16px;
-  outline: none;
-  background-color: transparent;
-  @include font(16px, 2vw, 20px, 400, 20px);
-  color: var(--text-color);
-
-  &:-webkit-autofill,
-  &:-webkit-autofill:focus {
-    transition: background-color 600000s 0s, color 600000s 0s;
-  }
-  &[data-autocompleted] {
-    background-color: transparent !important;
-  }
-
-  &.valid {
-    border-color: #33cc00;
-  }
-
-  &.invalid {
-    padding-right: 96px;
-    color: var(--text-color);
-    border-color: var(--invalid);
-  }
-  &:hover {
-    border-color: #33cc00;
-  }
-}
-
-.error {
-  position: absolute;
-  bottom: -18px;
-  left: 0;
-  color: var(--invalid);
-  @include font(10px, 2vw, 11px, 400, 12px);
-
-  @include mediaQuery($screen-size-m) {
-    @include font(12px, 2vw, 14px, 400, 14px);
-  }
-}
-
-.chip__eye {
-  position: absolute;
-  right: 4px;
-  top: 4px;
-  cursor: pointer;
-  fill: var(--text-color);
-}
-.warning + .chip__eye {
-  right: 49px;
-}
-
-.warning {
-  position: absolute;
-  top: 14px;
-  right: 14px;
-  width: 20px;
-  height: 20px;
-  border-radius: 50%;
-  background-color: var(--invalid);
-  @include flexCenter();
-  color: var(--text-color-dark);
-}
-
+// не можу знайти, де застосований цей класв
 .chip__checkbox__valid {
   position: absolute;
   top: 17px;
@@ -145,74 +11,4 @@
   border-radius: 50%;
   width: 21px;
   height: 21px;
-}
-.error__backend {
-  color: var(--invalid);
-  text-align: center;
-  @include font(14px, 2vw, 20px, 400, 19px);
-}
-
-.styledBtn {
-  @include flexCenter();
-  gap: 28px;
-  cursor: pointer;
-  background: #616161;
-  color: var(--text-color);
-
-  border-radius: var(--border-radius);
-  outline: none;
-  width: 280px;
-  margin: 0 auto;
-  @include font(14px, 2vw, 16px, 400, 19px);
-  padding: 10px 20px;
-  transition: all 0.3s;
-  text-transform: uppercase;
-  border: var(--border);
-
-  .warning {
-    position: static;
-    background-color: #fff;
-    color: var(--invalid);
-  }
-  &.invalid {
-    background-color: var(--invalid);
-    @include flexCenter();
-    gap: 8px;
-  }
-
-  &.valid {
-    background-color: var(--btn-background);
-
-    &:hover {
-      box-shadow: 0px 0px 8px #fff;
-    }
-  }
-}
-
-// linkedin + google
-.box__link__group {
-  padding-top: 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 24px;
-}
-.text {
-  @include font(14px, 2vw, 16px, 400, 19px);
-}
-.link__group {
-  display: flex;
-  align-items: center;
-  gap: 32px;
-}
-.link__icons {
-  cursor: pointer;
-}
-
-.link__forgot__password {
-  @include font(14px, 2vw, 16px, 400, 19px);
-
-  text-decoration: underline;
-  text-decoration-color: var(--text-color);
-  text-decoration-thickness: 1px;
 }
