+++ src/components/Auth/Register/register.module.scss
@@ -1,13 +1,7 @@
 @import '../../../sass/utils/container';
 @import '../../../sass/utils/mixins';
 
-@import '../Login/signIn.module.scss';
-
-.form {
-  display: flex;
-  flex-direction: column;
-}
-
+// RegisterFormik.tsx
 .form__box__checkbox {
   position: relative;
 
@@ -81,32 +75,5 @@
   &:hover {
     color: var(--hover-text-color);
     text-decoration-color: var(--hover-text-color);
-  }
-}
-
-.styledBtn {
-  margin: 0;
-  text-transform: uppercase;
-}
-
-.passwordStrength {
-  .strengthBar {
-    height: 3px;
-    border-radius: 3px;
-    width: 100%;
-    transition: background-color 0.3s ease;
-    margin-bottom: 4px;
-  }
-
-  .strengthLevels {
-    display: flex;
-    justify-content: space-between;
-    padding-right: 20px;
-    padding-left: 20px;
-
-    span {
-      transition: color 0.3s ease;
-      @include font(12px, 2vw, 14px, 400, 15px);
-    }
   }
 }
