+++ src/components/Auth/commonAuthStyles.module.scss
@@ -1,0 +1,100 @@
+@import '@/sass/utils/container';
+@import '@/sass/utils/mixins';
+
+.section {
+  padding: 40px 0;
+  @include mediaQuery($screen-size-s) {
+    padding: 74px 0;
+  }
+
+  @include mediaQuery($screen-size-l) {
+    padding: 94px 0;
+  }
+  @include mediaQuery($screen-size-xl) {
+    padding: 114px 0;
+  }
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+
+  padding-top: 52px;
+  margin: 0 auto;
+
+  max-width: 400px;
+  @include mediaQuery($screen-size-s) {
+    gap: 35px;
+  }
+  @include mediaQuery($screen-size-m) {
+    max-width: 530px;
+    gap: 48px;
+  }
+
+  @include mediaQuery($screen-size-xl) {
+    max-width: 584px;
+  }
+}
+
+.error__backend {
+  color: var(--invalid);
+  text-align: center;
+  @include font(14px, 2vw, 20px, 400, 19px);
+}
+
+//sunmit button
+.styledBtn {
+  @include flexCenter();
+  gap: 28px;
+  cursor: pointer;
+  background: #616161;
+  color: var(--text-color);
+
+  border-radius: var(--border-radius);
+  outline: none;
+  width: 280px;
+  margin: 0 auto;
+  @include font(14px, 2vw, 16px, 400, 19px);
+  padding: 10px 20px;
+  transition: all 0.3s;
+  text-transform: uppercase;
+  border: var(--border);
+
+  .warning {
+    position: static;
+    background-color: #fff;
+    color: var(--invalid);
+  }
+  &.invalid {
+    background-color: var(--invalid);
+    @include flexCenter();
+    gap: 8px;
+  }
+
+  &.valid {
+    background-color: var(--btn-background);
+
+    &:hover {
+      box-shadow: 0px 0px 8px #fff;
+    }
+  }
+}
+
+// social auth links container
+.box__link__group {
+  padding-top: 24px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 24px;
+}
+
+// забули пароль?
+.link__forgot__password {
+  @include font(14px, 2vw, 16px, 400, 19px);
+
+  text-decoration: underline;
+  text-decoration-color: var(--text-color);
+  text-decoration-thickness: 1px;
+}
