+++ src/components/UI/SocialLoginGroup/socialLoginGroup.module.scss
@@ -1,0 +1,13 @@
+@import '../../../sass/utils/mixins';
+
+.text {
+  @include font(14px, 2vw, 16px, 400, 19px);
+}
+.link__group {
+  display: flex;
+  align-items: center;
+  gap: 32px;
+}
+.link__icons {
+  cursor: pointer;
+}
